refactor: Move Ui.java to ui package
