+++ src/main/java/zbot/ui/MainWindow.java
@@ -7,7 +7,6 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import zbot.Ui;
 import zbot.ZBot;
 
 /**
