+++ src/main/java/zbot/ui/Ui.java
@@ -1,4 +1,4 @@
-package zbot;
+package zbot.ui;
 
 import java.util.Scanner;
